requirements.txt
transformers
bitsandbytes
accelerate
PyPDF2
python-dotenv